Refactor: 모킹 api 서버 json-server에서 express 경량 서버로 변경 (#14)
Resolves: #13
