+++ mocks/routes/user-service.mjs
@@ -1,7 +1,6 @@
-const express = require("express"); 
+import express from "express";
+import token from "../data/auth/auth.json" with { type: "json" };
 const router = express.Router();
-
-const token = require("../data/auth/auth.json")
 
 router.post("/auth/login", (req, res) => {
   res.json(token);
@@ -9,10 +8,10 @@ router.post("/auth/login", (req, res) => {
 
 router.post("/auth/logout", (req, res) => {
   res.sendStatus(204);
-})
+});
 
 router.post("/auth/refresh", (req, res) => {
   res.json(token);
-})
+});
 
-module.exports = router;
+export default router;
