+++ package.json
@@ -6,7 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "mock": "node mocks/server.js",
+    "mock": "node mocks/server.mjs",
     "lint": "eslint",
     "prepare": "husky install",
     "test": "vitest",
@@ -40,6 +40,7 @@
     "@vitejs/plugin-react": "^5.1.4",
     "@vitest/ui": "^3.2.4",
     "autoprefixer": "^10.4.27",
+    "cors": "^2.8.6",
     "eslint": "^9",
     "eslint-config-next": "16.1.6",
     "eslint-config-prettier": "^10.1.8",
@@ -47,7 +48,6 @@
     "express": "^5.2.1",
     "husky": "^8.0.0",
     "jsdom": "^27.0.1",
-    "json-server": "^0.17.4",
     "postcss": "^8.5.8",
     "prettier": "^3.8.1",
     "tailwindcss": "^4.2.1",
